% Auto-generated: do not edit by hand
\name{threeJsOrientation}

\alias{threeJsOrientation}

\title{ThreeJsOrientation component}

\description{

}

\usage{
threeJsOrientation(id=NULL, activeTime=NULL, cameraFollowModel=NULL, data=NULL,
headingOffset=NULL, headingSign=NULL, modelFile=NULL,
pitchOffset=NULL, pitchSign=NULL, rollOffset=NULL,
rollSign=NULL, rotationOrder=NULL, style=NULL,
textureFile=NULL)
}

\arguments{
\item{id}{Character. }

\item{activeTime}{Numeric. }

\item{cameraFollowModel}{Logical. }

\item{data}{Character. }

\item{headingOffset}{Numeric. }

\item{headingSign}{Numeric. }

\item{modelFile}{Character. }

\item{pitchOffset}{Numeric. }

\item{pitchSign}{Numeric. }

\item{rollOffset}{Numeric. }

\item{rollSign}{Numeric. }

\item{rotationOrder}{List of characters. }

\item{style}{Named list. }

\item{textureFile}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

