% Auto-generated: do not edit by hand
\name{threeJsOrientation}

\alias{threeJsOrientation}

\title{ThreeJsOrientation component}

\description{

}

\usage{
threeJsOrientation(id=NULL, activeTime=NULL, data=NULL, fbxFile=NULL,
style=NULL)
}

\arguments{
\item{id}{Character. }

\item{activeTime}{Character. }

\item{data}{Character. }

\item{fbxFile}{Character. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

