% Auto-generated: do not edit by hand
\name{threeJsOrientation}

\alias{threeJsOrientation}

\title{ThreeJsOrientation component}

\description{

}

\usage{
threeJsOrientation(id=NULL, activeTime=NULL, data=NULL, objFile=NULL,
style=NULL, textureFile=NULL)
}

\arguments{
\item{id}{Character. }

\item{activeTime}{Numeric. }

\item{data}{Character. }

\item{objFile}{Character. }

\item{style}{Named list. }

\item{textureFile}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

